+++ src/app/[locale]/posts/[...slug]/page.tsx
@@ -1,6 +1,9 @@
 'use server';
 
+import { notFound } from 'next/navigation';
+
 import MDComponent from '@/entities/markdown';
+import Reply from '@/shared/components/reply';
 
 interface PostDetailPageProps {
   params: Promise<{ locale: LocaleType; slug: string[] }>;
@@ -9,6 +12,16 @@ interface PostDetailPageProps {
 export default async function PostDetailPage(props: PostDetailPageProps) {
   const { locale, slug } = await props.params;
 
-  const path = [locale, 'posts', ...slug].join('/') + '.mdx';
-  return <MDComponent path={path} />;
+  if (!slug || slug.length === 0) {
+    return notFound();
+  }
+
+  const path = [locale, ...slug].join('/') + '.mdx';
+
+  return (
+    <>
+      <MDComponent path={path} />
+      <Reply locale={locale} />
+    </>
+  );
 }