+++ src/features/post/ui/post-basic-card.tsx
@@ -1,19 +1,26 @@
 'use client';
 
+import Image from 'next/image';
+
 import { Frontmatter } from '@/entities/markdown/model/markdown.schema';
-import { Button } from '@/shared/components/ui/button';
+import Button from '@/shared/components/ui/button';
+import Link from '@/shared/components/ui/link';
+
 import { format } from 'date-fns';
 import { Link2 } from 'lucide-react';
-import Image from 'next/image';
-import Link from 'next/link';
+
+interface Props extends Frontmatter {
+  locale: string;
+}
 
 export default function PostBasicCard({
+  locale,
   title,
   thumbnail,
   createdAt,
   path,
   summary,
-}: Frontmatter) {
+}: Props) {
   return (
     <article className="grid grid-cols-12 gap-y-4 gap-4 p-6 rounded-lg border border-zinc-100 bg-white shadow dark:border-zinc-800 dark:bg-gray-800">
       <div className="col-span-full md:col-span-6">
@@ -36,6 +43,7 @@ export default function PostBasicCard({
           <Button
             variant="primary"
             className="place-self-end text-sm"
+            nativeButton={false}
             render={<Link href={['/posts', ...path].join('/')} />}
           >
             <Link2 size={16} />